src/darkdawn/requirements/base.txt
django==3.2.*
django-ipware